+++ desktop/qml/TaskTimegraph.qml
@@ -9,47 +9,55 @@ Item {
     id: control
 
     property int taskId: -1
-    property string type: Task.type(taskId)
-    property string taskColor: Task.color(taskId)
-    property int duration: Task.duration(taskId)
-    property date assignedDate: Task.assignedDate(taskId)
-    property date taskEndDate: MDate.addDays(assignedDate, duration)
+    property string taskColor: taskMap["color"]
+    property int duration: taskMap["duration"]
     property date seasonBegin
+    property int season
+    property int year
+    property var taskMap: Task.drawInfoMap(taskId, season, year)
+    property string description: taskId > 0 ? taskMap["description"] : ""
+    property alias hovered: mouseArea.containsMouse
 
+    x: taskMap["graphStart"]
     height: Units.rowHeight
     implicitHeight: Units.rowHeight
-    width: taskBar.x + taskBar.width
+    width: taskBar.width + (label.wideEnough ? 0 : label.implicitWidth)
 
-    function acronymize(string) {
-        var stringList = string.split(" ");
-        if (stringList.length > 1) {
-            var s = ""
-            for (var i = 0; i < stringList.length; i++)
-                s += stringList[i][0].toString().toUpperCase();
-            return s;
-        } else {
-            return stringList[0][0] + stringList[0][1].toString().toUpperCase();
-        }
+    ToolTip.visible: mouseArea.containsMouse
+    ToolTip.delay: 200
+    ToolTip.text: taskMap["toolTip"]
+
+    MouseArea {
+        id: mouseArea
+        anchors.fill: parent
+        hoverEnabled: true
     }
 
     Rectangle {
         id: taskBar
-        x: Units.position(seasonBegin, assignedDate)
-        width: Units.widthBetween(x, seasonBegin, taskEndDate)
-        visible: width > 0
+        width: taskMap["width"]
+        radius: 3
+        visible: duration > 0 && width > 0
         height: parent.height * 0.7
         anchors.verticalCenter: parent.verticalCenter
         color: taskColor
+        opacity: 0.8
 
         Label {
-            text: acronymize(type)
+            id: label
+            property bool wideEnough: parent.width > 20
+            anchors {
+                left: parent.left
+                verticalCenter: parent.verticalCenter
+                leftMargin: wideEnough ? 4 : 0
+            }
+            text: description
+            visible: taskBar.width
+            rotation: wideEnough ? 0 : -90
             font.family: "Roboto Condensed"
             font.pixelSize: Units.fontSizeBodyAndButton
             antialiasing: true
-            color: Material.color(Material.Grey, Material.Shade100)
-            anchors.left: parent.left
-            anchors.verticalCenter: parent.verticalCenter
-            anchors.leftMargin: 4
+            color: Material.color(Material.Grey, taskBar.width < 10 ? Material.Shade600 : Material.Shade100)
         }
     }
 }